requirements.txt
python-dotenv
python-multipart
pymongo[srv]
fastapi
uvicorn
pytest
httpx